implemented camera

+++ src/dood.rs
@@ -56,8 +56,8 @@ impl Entity for Dood {
 
     fn render(&self, tiles: &Pixset) -> Vec<Vertex> {
         let offset = (SQUARE_SIZE / 2) as f32;
-        let x = (self.loc.0 * SQUARE_SIZE as u32) as f32 + offset;
-        let y = (self.loc.1 * SQUARE_SIZE as u32) as f32 + offset;
+        let x = (self.loc.0 * SQUARE_SIZE) as f32 + offset;
+        let y = (self.loc.1 * SQUARE_SIZE) as f32 + offset;
 
         return vec![
             // bottom left
